requirements.txt
altgraph==0.17.4
packaging==24.1
pefile==2024.8.26
pillow==11.0.0
ping3==4.0.8
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pywin32-ctypes==0.2.3
